Fixed processing timeout issue
This actually involves 2 bugs/fixes.

The first is that the tick clock is not reset properly at the start
of syncPPM function. This means that sometimes we enter the function
with tick already more than BLANK_CHECK interval. In which case we
exit immediately instead of waiting for sync.

The second is that because the input pulse train changes in length
the remaining time may or may not be exactly 11ms. To remedy that
the timeout is now calculated from the start of the input pulse
train and set to 19ms.