Implemented input auto-calibration at startup
This is to simplify calculations for trim etc since we can be sure
that the input is 0 (or very close to it) when the stick is at neutral.
The downside is that we must always turn the TX on with the trims at
neutral to avoid false centering.

We ignore channel 3 (throttle) when doing this because it doesn't have a
real neutral position since it doesn't have springs.